Removes geofence code for sdk<26
