+++ app/src/main/AndroidManifest.xml
@@ -70,20 +70,6 @@
             android:enabled="true"
             android:permission="android.permission.BIND_JOB_SERVICE" />
 
-        <service
-            android:name="com.visilabs.gps.geofence.GeofenceMonitor"
-            android:enabled="true"
-            android:exported="true" />
-
-        <receiver
-            android:name="com.visilabs.gps.geofence.GeofenceTransitionsReceiver"
-            android:enabled="true"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="com.visilabs.android.gps.geofence.ACTION_RECEIVE_GEOFENCE" />
-            </intent-filter>
-        </receiver>
-
          <receiver
             android:name="com.visilabs.gps.geofence.VisilabsAlarm"
             android:exported="false" />
