+++ visilabs-android/src/main/java/com/visilabs/Injector.java
@@ -2,13 +2,11 @@ package com.visilabs;
 
 
 import com.visilabs.gps.manager.GpsManager;
-import com.visilabs.gps.manager.GpsManagerMoreThanOreo;
 
 public enum Injector {
     INSTANCE;
 
     private GpsManager gpsManager;
-    private GpsManagerMoreThanOreo gpsManagerMoreThanOreo;
 
     Injector(){
     }
@@ -17,15 +15,7 @@ public enum Injector {
         this.gpsManager = gpsManager;
     }
 
-    public void initGpsManager(GpsManagerMoreThanOreo gpsManagerMoreThanOreo) {
-        this.gpsManagerMoreThanOreo = gpsManagerMoreThanOreo;
-    }
-
     public GpsManager getGpsManager() {
         return gpsManager;
-    }
-
-    public GpsManagerMoreThanOreo getGpsManagerMoreThanOreo() {
-        return gpsManagerMoreThanOreo;
     }
 }
