--- visilabs-android/src/main/java/com/visilabs/gps/listener/IVisilabsGeofenceListener.java
@@ -1,11 +1,0 @@
-package com.visilabs.gps.listener;
-
-import com.google.android.gms.location.Geofence;
-
-public interface IVisilabsGeofenceListener {
-    void onEnter(Geofence geofence);
-
-    void onExit(Geofence geofence);
-
-    void onTrigger(Geofence geofence);
-}
