+++ visilabs-android/src/main/java/com/visilabs/gps/util/GeoFencesUtils.java
@@ -4,19 +4,8 @@ import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
 
-import com.visilabs.gps.geofence.GeofenceTransitionsReceiver;
-
 
 public class GeoFencesUtils {
-    public static PendingIntent getTransitionPendingIntent(Context context) {
-        Intent intent = new Intent(context, GeofenceTransitionsReceiver.class);
-        //Intent intent = new Intent("com.visilabs.gps.geofence.ACTION_RECEIVE_GEOFENCE");
-        return PendingIntent.getBroadcast(
-                context,
-                0,
-                intent,
-                PendingIntent.FLAG_UPDATE_CURRENT);
-    }
 
     private static final double R = 6372.8; // In kilometers
